+++ addons/external.go
@@ -94,4 +94,11 @@ func init() {
 		},
 		Module: "github.com/invopop/gobl.dk.oioubl",
 	})
+	tax.RegisterApprovedAddon(&tax.ExternalAddon{
+		Key: "fi-finvoice-v3",
+		Name: i18n.String{
+			i18n.EN: "Finland Finvoice 3.0",
+		},
+		Module: "github.com/invopop/gobl.fi.finvoice",
+	})
 }